Recursos/Backend/CamejobsBackEnd/requirements.txt
django
djangorestframework
mysqlclient
Crear la base de datos en el puerto 33065 con el nombre de camejobs, cambiar el user y password dentro del settings por
el de ustedes, después de que la base de datos y el user-password estén bien asimilados, realizar migraciones
python manage.py makemigrations